// lib/dxp/types/dxp-api.d.ts
export interface components {
    schemas: {
// ...
        BcicLeadCreateRequest: {
            addresses: components["schemas"]["BcicLeadCreateRequestAddress"][];
            agencies?: components["schemas"]["LeadRequestAgency"][];
            chats?: components["schemas"]["LeadCreateRequestChat"][];
            /** @example BB */
            citizenshipCd: string;
            /**
             * Format: date
             * @example 1985-07-13
             */
            dateOfBirth: string;
            emails?: components["schemas"]["LeadCreateRequestEmail"][];
            employer?: string;
            /** @example EMP_FT */
            employmentStatusCd: string;
            /** @example John */
            firstName: string;
            /**
             * @description use 'GenderCd' lookup code
             * @example male
             */
            gender?: string;
            /** @example NID */
            identificationTypeCd: string;
            /** @example Jones */
            lastName: string;
            /** @example Steve */
            middleName?: string;
            /** @example ITSpecTech */
            occupationCd: string;
            /** @example GOP */
            organizationName?: string;
            phones?: components["schemas"]["LeadCreateRequestPhone"][];
            /** @example false */
            seniorPublicOfficeInd?: boolean;
            socialNets?: components["schemas"]["LeadCreateRequestSocialNet"][];
            /**
             * @description use 'BcicSourceOfFund' lookup code
             * @example Salary
             */
            sourceOfFundCd: string;
            /** @example Investments */
            sourceOfFundDetails?: string;
            /** @example 2AAA */
            taxId: string;
            /** @example Party President */
            titlePEP?: string;
            /** @example 123456789 */
            trn?: string;
            webAddresses?: components["schemas"]["LeadCreateRequestWebAddress"][];
        };
// ...
        Email: {
            /** @example email@email.com */
            email: string;
        };
// ...
        LeadCreationResult: {
            customerNumber?: string;
        };
// ...
    };
// ...
}